is-lnu-service/src/main/java/org/lnu/is/security/LogoutSuccessHandler.java
package org.lnu.is.security;

import org.springframework.security.core.Authentication;
import org.springframework.security.web.authentication.logout.SimpleUrlLogoutSuccessHandler;

import javax.servlet.ServletException;
import javax.servlet.http.HttpServletRequest;
import javax.servlet.http.HttpServletResponse;
import java.io.IOException;

/**
 * Logout Success Handler.
 * Needed for closing session of edbo service.
 * @author ivanursul
 *
 */
public class LogoutSuccessHandler extends SimpleUrlLogoutSuccessHandler {

	@Override
	public void onLogoutSuccess(final HttpServletRequest request, final HttpServletResponse response, final Authentication authentication) throws IOException, ServletException {
		super.onLogoutSuccess(request, response, authentication);
	}

}
